Refactor inference-logger to be in api directory
