+++ api/cmd/inference-logger/main_test.go
