+++ api/pkg/inference-logger/logger/console_sink.go
