+++ api/pkg/inference-logger/mocks/NewRelicLogsClient.go
