+++ infra/kafka/Makefile
@@ -1,0 +1,3 @@
+.PHONY: setup-kafka
+setup-kafka:
+	@docker-compose up -d
